requirements.txt
crewai==1.9.0
crewai-tools==1.9.0
streamlit==1.53.1
python-dotenv==1.1.1
litellm==1.9.4
litellm[proxy]
pyyaml==6.0.1
pydantic>=2.0.0